+++ src/walking/KeyInput.java
@@ -43,7 +43,7 @@ public class KeyInput {
     }
 
     void mousClick(MouseEvent e, double shootDirectionX, double shootDirectionY) {
-        // NIEUWE bullet object aanmaken van locatie PLAYER
-        handler.addObject(new Bullet(handler.getGameObject(ID.Player).getX(), handler.getGameObject(ID.Player).getY(), ID.Bullet, handler, e.getSceneX(), e.getSceneY()));
+        // NIEUW bullet object aanmaken van locatie PLAYER
+        handler.addObject(new Bullet(handler.getGameObject(ID.Player).getX(), handler.getGameObject(ID.Player).getY(), ID.Bullet, handler, (int)e.getSceneX(), (int)e.getSceneY()));
     }
 }
